feat: evaluate player guess and provide feedback
This push introduces the input_guess method which prompts the human
player for a guess when the human plays as the Code Breaker.

The eval_guess method then compares the guess against the computer's
master code to determine if the guess contains any color pegs that are
exact matches (correct color and order) or partial matches (correct
color but in the wrong order).

The update_guess_row, update_key_row, and print_board methods print the
guess on the game board along with key pegs that provide feedback to
the human player of how many exact/partial matches are present within
the guess.